✨ #4 configファイルの削除、ファイル移動による修正、`psycopg2`を用いたPosgres用のセットアップ
