--- server/app/config.py
@@ -1,16 +1,0 @@
-import os
-
-
-# TODO: データベース関連の設定を後にPostgreSQLに変更する
-class Config:
-    SECRET_KEY = os.environ.get("SECRET_KEY") or "you-will-never-guess"
-    SQLALCHEMY_DATABASE_URI = os.environ.get("DATABASE_URL") or "sqlite:///app.db"
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-
-class DevelopmentConfig(Config):
-    DEBUG = True
-
-
-class ProductionConfig(Config):
-    DEBUG = False
